samples: bluetooth: channel_sounding: factor out shared sample code
Move shared CS sample helpers (connection/scan callbacks, CS event
logging, and median utilities) into a common module and update the
IPT/RAS initiator and reflector samples to use it. This reduces
duplicated code and keeps sample behavior consistent when shared logic
is updated.
