+++ samples/bluetooth/channel_sounding/ipt_reflector/README.rst
@@ -78,22 +78,52 @@ After programming the sample to your development kit, test it together with a de
 #. The reflector starts advertising and the initiator scans for ``Nordic CS IPT Reflector``.
    Once the initiator connects, check the reflector terminal for information similar to the following::
 
-      *** Booting nRF Connect SDK v3.2.99-09f8a95ee5b4 ***
-      *** Using Zephyr OS v4.3.99-2925b3840984 ***
+      *** Booting nRF Connect SDK v3.2.99-4dc52e0d80a6 ***
+      *** Using Zephyr OS v4.4.0-2b3d25313ec3 ***
       I: Starting Channel Sounding IPT Reflector Sample
       I: SoftDevice Controller build revision:
-      I: 36 9c e5 98 e5 71 0a a0 |6....q..
-      I: 2d b9 89 96 6b 15 65 4b |-...k.eK
-      I: d6 e8 4a ac             |..J.
+      I: 10 7f 81 88 e0 eb fc fc |........
+      I: 53 3c 0e d3 0f e7 d0 4f |S<.....O
+      I: 0c 64 24 fc             |.d$.
       I: HW Platform: Nordic Semiconductor (0x0002)
       I: HW Variant: nRF54Lx (0x0005)
-      I: Firmware: Standard Bluetooth controller (0x00) Version 54.58780 Build 175236504
+      I: Firmware: Standard Bluetooth controller (0x00) Version 16.33151 Build 4243316872
       I: HCI transport: SDC
       I: Identity: D2:CE:07:7E:40:79 (random)
-      I: HCI: version 6.2 (0x10) revision 0x30a9, manufacturer 0x0059
-      I: LMP: version 6.2 (0x10) subver 0x30a9
+      I: HCI: version 6.2 (0x10) revision 0x30b8, manufacturer 0x0059
+      I: LMP: version 6.2 (0x10) subver 0x30b8
       I: Connected to E6:BA:C9:31:51:DB (random) (err 0x00)
       I: CS capability exchange completed.
+      Remote CS capabilities:
+      - num_config_supported: 1
+      - max_consecutive_procedures_supported: 0
+      - num_antennas_supported: 1
+      - max_antenna_paths_supported: 1
+      - initiator_supported: yes
+      - reflector_supported: no
+      - mode_3_supported: no
+      - rtt_aa_only_precision: 150 ns (2)
+      - rtt_sounding_precision: not supported (0)
+      - rtt_random_payload_precision: 150 ns (2)
+      - rtt_aa_only_n: 30
+      - rtt_sounding_n: 0
+      - rtt_random_payload_n: 30
+      - phase_based_nadm_sounding_supported: no
+      - phase_based_nadm_random_supported: no
+      - cs_sync_2m_phy_supported: yes
+      - cs_sync_2m_2bt_phy_supported: no
+      - cs_without_fae_supported: yes
+      - chsel_alg_3c_supported: no
+      - pbr_from_rtt_sounding_seq_supported: no
+      - t_ip1_times_supported: 0x007c
+      - t_ip2_times_supported: 0x007e
+      - t_fcs_times_supported: 0x01e0
+      - t_pm_times_supported: 0x0003
+      - t_sw_time: 0 us
+      - tx_snr_capability: 0x00
+      - t_ip2_ipt_times_supported: 0x007e
+      - t_sw_ipt_time_supported: 10 us
+
       I: CS config creation complete.
       - id: 0
       - mode: 2 (PBR)
